Revert "More precision for duration and latency"
This reverts commit f44bbed54b0ffa3ee575ad3c1a20ed0baa931704.

Conflicts:
	src/Hangfire.Core/Dashboard/HtmlHelper.cs
	src/Hangfire.Core/Dashboard/JobHistoryRenderer.cs
	src/Hangfire.Core/Dashboard/Pages/SucceededJobs.cshtml
	src/Hangfire.Core/Dashboard/Pages/SucceededJobs1.generated.cs
	src/Hangfire.SqlServer/SqlServerMonitoringApi.cs

+++ src/Hangfire.Core/States/SucceededState.cs
@@ -26,7 +26,7 @@ namespace Hangfire.States
     {
         public static readonly string StateName = "Succeeded";
 
-        public SucceededState(object result, double latency, double performanceDuration)
+        public SucceededState(object result, long latency, long performanceDuration)
         {
             Result = result;
             Latency = latency;
@@ -36,8 +36,8 @@ namespace Hangfire.States
 
         public object Result { get; private set; }
         public DateTime SucceededAt { get; private set; }
-        public double Latency { get; private set; }
-        public double PerformanceDuration { get; private set; }
+        public long Latency { get; private set; }
+        public long PerformanceDuration { get; private set; }
 
         public string Name { get { return StateName; } }
         public string Reason { get; set; }
